requirements.txt
spotipy==2.24.0
python-dotenv==1.0.1
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1